Refactor project to be more deployable
