+++ pyproject.toml
@@ -7,6 +7,8 @@ requires-python = ">=3.13"
 dependencies = [
     "boto3>=1.35.50",
     "flask>=3.0.3",
+    "gunicorn>=23.0.0",
+    "setproctitle>=1.3.3",
 ]
 
 [tool.uv]
@@ -16,7 +18,7 @@ dev-dependencies = [
 ]
 
 [tool.pytest.ini_options]
-pythonpath = "src"
+pythonpath = [ ".", "src"]
 testpaths = [
     "tests"
 ]
@@ -36,7 +38,6 @@ select = [
     # isort
     "I",
 ]
-
 
 [tool.ruff.format]
 quote-style = "double"
