+++ src/__init__.py
@@ -1,0 +1,4 @@
+from app import create_app
+
+if __name__ == "__main__":
+    create_app().run(port=8080)
