--- src/disk_reader.py
@@ -1,10 +1,0 @@
-import game
-
-
-def read_text(path: str) -> game.TilePool:
-    tiles = []
-    with open(path) as f:
-        for line in f:
-            tiles.append(game.Tile(line.strip()))
-
-    return game.TilePool(frozenset(tiles))
